# doc/openapi.yaml
openapi: 3.0.3
info:
  title: 应用发布服务 API
  version: 1.0.0
  description: 当前应用发布服务接口的 OpenAPI 定义。
servers:
  - url: http://localhost:8080
    description: 本地环境
...
paths:
...
  /api/store-configs:
    get:
      tags: [渠道配置]
      summary: 查询渠道配置列表
      parameters:
        - in: query
          name: current
          description: 当前页，兼容 pageNum
          schema:
            type: integer
            format: int64
        - in: query
          name: size
          description: 每页大小，兼容 pageSize
          schema:
            type: integer
            format: int64
        - in: query
          name: pageNum
          description: 页码
          schema:
            type: integer
            format: int64
        - in: query
          name: pageSize
          description: 每页条数
          schema:
            type: integer
            format: int64
        - in: query
          name: key
          description: 渠道名称或账号关键字
          schema:
            type: string
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponsePageResponseStoreConfigResponse'
    post:
      tags: [渠道配置]
      summary: 创建渠道配置
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StoreConfigRequest'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseStoreConfigResponse'
  /api/store-configs/{configId}:
    get:
      tags: [渠道配置]
      summary: 查询渠道配置详情
      parameters:
        - $ref: '#/components/parameters/ConfigId'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseStoreConfigResponse'
    put:
      tags: [渠道配置]
      summary: 更新渠道配置
      parameters:
        - $ref: '#/components/parameters/ConfigId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StoreConfigRequest'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseStoreConfigResponse'
    delete:
      tags: [渠道配置]
      summary: 删除渠道配置
      parameters:
        - $ref: '#/components/parameters/ConfigId'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
  /api/store-configs/{configId}/status:
    put:
      tags: [渠道配置]
      summary: 更新渠道配置状态
      parameters:
        - $ref: '#/components/parameters/ConfigId'
        - in: query
          name: apiStatus
          required: true
          schema:
            type: integer
            enum: [0, 1]
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseStoreConfigResponse'
...
components:
  parameters:
...
    ConfigId:
      in: path
      name: configId
      description: 渠道配置 ID
      required: true
      schema:
        type: integer
        format: int64
...
  schemas:
...
    StoreConfigRequest:
      type: object
      required: [storeType]
      properties:
        id:
          type: integer
          format: int64
        storeType:
          type: string
          description: 渠道类型编码
        accountName:
          type: string
        email:
          type: string
        phone:
          type: string
        clientId:
          type: string
        clientSecret:
          type: string
        miPublicKey:
          type: string
        miPrivateKey:
          type: string
        token:
          type: string
        ipWhitelist:
          type: string
        apiStatus:
          type: integer
          enum: [0, 1]
          description: 接口对接状态，0=关闭，1=开启
...
    StoreConfigResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        storeType:
          type: string
          description: 渠道类型名称
        accountName:
          type: string
        email:
          type: string
        phone:
          type: string
        clientId:
          type: string
        clientSecret:
          type: string
        miPublicKey:
          type: string
        miPrivateKey:
          type: string
        token:
          type: string
        ipWhitelist:
          type: string
        apiStatus:
          type: integer
          description: 接口对接状态，0=关闭，1=开启
        createUser:
          type: string
        updateUser:
          type: string
        createTime:
          type: string
          format: date-time
        updateTime:
          type: string
          format: date-time
    PageResponseStoreConfigResponse:
      type: object
      properties:
        current:
          type: integer
          format: int64
        size:
          type: integer
          format: int64
        total:
          type: integer
          format: int64
        pages:
          type: integer
          format: int64
        records:
          type: array
          items:
            $ref: '#/components/schemas/StoreConfigResponse'
...
    ApiResponseStoreConfigResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/StoreConfigResponse'
        message:
          type: string
    ApiResponsePageResponseStoreConfigResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/PageResponseStoreConfigResponse'
        message:
          type: string
...
    ApiResponseVoid:
      type: object
      properties:
        success:
          type: boolean
        data:
          nullable: true
        message:
          type: string